requirements.txt
django==5.2
environs[django]==14.1.1
djangorestframework==3.16.0
gunicorn==23.0.0
psycopg[binary]==3.2.6
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
drf_yasg==1.21.10
django_filter==25.1